fixes en delete_project

+++ delete_project.sh
@@ -23,6 +23,9 @@ ask_folder_to_install_sites
 MENU_TITLE="PROJECT TO DELETE"
 directory_browser "${MENU_TITLE}" "${FOLDER_TO_INSTALL}"
 
+# Creating a tmp directory
+mkdir ${SFOLDER}/tmp-backup
+
 echo "directory_broser returns: " $filepath"/"$filename
 if [[ -z "${filepath}" || ${filepath} == "" ]]; then
 
@@ -30,27 +33,37 @@ if [[ -z "${filepath}" || ${filepath} == "" ]]; then
 
 else
 
-    # Creating a tmp directory
-    mkdir ${SFOLDER}/tmp-backup
-
-    # Making a backup of project files
-    echo -e ${CYAN}" > Making a backup ..."${ENDCOLOR}
-    cp -r $filepath"/"$filename ${SFOLDER}/tmp-backup
-    echo -e ${GREEN}" > Project files stored: ${SFOLDER}/tmp-backup"${ENDCOLOR}
-
-    # Deleting project files
-    rm -R $filepath"/"$filename
-    echo -e ${GREEN}" > Project Files Deleted!"${ENDCOLOR}
-
     # Removing last slash from string
     filename=${filename%/}
 
-    # Making a copy of nginx configuration file
-    cp -r /etc/nginx/sites-available/${filename} ${SFOLDER}/tmp-backup
+    # Making a backup of project files
+    echo -e ${CYAN}" > Making a backup ..."${ENDCOLOR}
+    #cp -r $filepath"/"$filename ${SFOLDER}/tmp-backup
+    TAR_FILE=$($TAR --exclude '.git' --exclude '*.log' -jcpf ${SFOLDER}/tmp-backup/backup-${filename}_files.tar.bz2 --directory=${FOLDER_TO_INSTALL} ${filename} >>$LOG)
 
-    # Deleting nginx configuration file
-    rm /etc/nginx/sites-available/${filename}
-    rm /etc/nginx/sites-enabled/${filename}
+    if ${TAR_FILE}; then
+
+        echo -e ${GREEN}" > Backup project files stored: ${SFOLDER}/tmp-backup"${ENDCOLOR}
+
+        echo " > Trying to create folder ${FOLDER_NAME} in Dropbox ..." >>$LOG
+        OLD_SITES_DP_F="/old-sites"
+        ${DPU_F}/dropbox_uploader.sh mkdir /${OLD_SITES_DP_F}
+        ${DPU_F}/dropbox_uploader.sh mkdir /${OLD_SITES_DP_F}/${filename}/
+
+        echo " > Uploading backup to Dropbox ..." >>$LOG
+        ${DPU_F}/dropbox_uploader.sh upload ${SFOLDER}/tmp-backup/backup-${filename}_files.tar.bz2 ${OLD_SITES_DP_F}/${filename}/
+
+        # Deleting project files
+        #rm -R $filepath"/"$filename
+        echo -e ${GREEN}" > Project files deleted from ${FOLDER_TO_INSTALL}!"${ENDCOLOR}
+
+        # Making a copy of nginx configuration file
+        cp -r /etc/nginx/sites-available/${filename} ${SFOLDER}/tmp-backup
+
+        # Deleting nginx configuration file
+        rm /etc/nginx/sites-available/${filename}
+        rm /etc/nginx/sites-enabled/${filename}
+    fi
 
 fi
 
@@ -67,13 +80,17 @@ if [ $exitstatus = 0 ]; then
     USER_DB="${PROJECT_NAME}_user"
 
     # Making a database Backup
-    mysql_database_export "${CHOSEN_DB}" "${filename}_DB.sql"
+    mysql_database_export "${CHOSEN_DB}" "${SFOLDER}/tmp-backup/${CHOSEN_DB}_DB.sql"
 
-    # Deleting mysql user
-    mysql_user_delete "${USER_DB}"
-
+    # TODO: TAR Backup and Upload to Dropbox
+    
     # Deleting project database
     mysql_database_drop "${CHOSEN_DB}"
+
+    # Deleting mysql user
+
+    # TODO: need to ask
+    #mysql_user_delete "${USER_DB}"
 
 else
     exit 1
